pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pokewatch"]

[project]
name = "pokewatch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0.0",
    "pyyaml",
    "python-dotenv",
    "requests",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=12.0.0",
    "pytest",
    "pytest-mock",
    "pytest-cov",
    "httpx",
    "mlflow>=2.10.0",
    "matplotlib",
    "dvc[s3]",
    "bentoml>=1.2.0",
    "zenml[server]>=0.55.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "black",
    "mypy",
    "pre-commit",
]
